fix: undoのupdate
