+++ src/editor/ViewPort/MainViewItems/Objects.tsx
@@ -118,7 +118,7 @@ const StaticObject = ({ om }) => {
       // animationsもコピー
       clone.animations = animations;
       if (id) {
-        setArg(id, "animations", animations);
+        setArg(id, "animations", animations, false);
       }
       if (om.args.castShadow) {
         clone.traverse((node) => {
